Use ES6 modules instead of commonJs
When using a Top-Level Await, webpack isn't aware that the script is
a module, so it throws an error for trying to use a Top-Level Await in
a script that isn't a module. To fix that, i've added "type: module"
to package.json and standardized the entire boilerplate to use ES6
modules.